requirements.txt
# FounderForge AI Cofounder Dependencies
# Core AI and workflow dependencies
langgraph>=0.0.40
google-generativeai>=0.3.0
sentence-transformers>=2.2.0

# Local storage and search
faiss-cpu>=1.7.0

# Web interface
streamlit>=1.28.0

# Testing and monitoring
pytest>=7.4.0
psutil>=5.9.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0